+++ src/main/java/lotto/Main.java
@@ -1,6 +1,6 @@
 package lotto;
 
-import lotto.domain2.*;
+import lotto.domain.*;
 import lotto.view.InputView;
 import lotto.view.ResultView;
