+++ src/main/java/lotto/domain/Lotto.java
@@ -1,37 +1,39 @@
 package lotto.domain;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import java.util.stream.Collectors;
 
 public class Lotto {
+    static final int NUMBER_COUNT = 6;
+    private final List<LottoNumber> lottoNumbers;
+    private final boolean isDirect;
 
-    private final List<LottoNo> numbers;
-
-    public Lotto(final List<Integer> numbers) {
-        final List<LottoNo> lottoNos = new ArrayList<>();
-        for (Integer number : numbers) {
-            lottoNos.add(new LottoNo(number));
-        }
-        this.numbers = lottoNos;
+    public Lotto(final List<LottoNumber> lottoNumbers) {
+        this(lottoNumbers, false);
     }
 
-    Rank checkRank(final List<Integer> winNumbers, final int bonusNumber) {
-        int countOfMatch = (int) numbers.stream().filter(number -> number.contains(winNumbers)).count();
-        final boolean isMatchBonus = numbers.stream().anyMatch(number -> number.getNumber() == bonusNumber);
-
-        if (countOfMatch == Rank.FOURTH.getCountOfMatch() && isMatchBonus) {
-            countOfMatch = countOfMatch + 1;
+    public Lotto(final List<LottoNumber> numbers, final boolean isDirect) {
+        if (NUMBER_COUNT != numbers.size()) {
+            throw new IllegalArgumentException(String.format("로또번호의 개수는 %s개만 가능합니다. 입력받은 로또번호 개수 = %s", NUMBER_COUNT, numbers.size()));
         }
 
-        return Rank.findByCountOfMatch(countOfMatch, isMatchBonus);
+        this.lottoNumbers = numbers;
+        this.isDirect = isDirect;
     }
 
-    public List<Integer> findNumbers() {
-        return this.numbers.stream()
-                .map(LottoNo::getNumber)
-                .collect(Collectors.toList());
+    int getCountOfMatchWith(final List<LottoNumber> winningLotto) {
+        return (int) lottoNumbers.stream()
+                .filter(winningLotto::contains)
+                .count();
+    }
+
+    boolean isMatchBonus(final int bonusNumber) {
+        return lottoNumbers.stream()
+                .anyMatch(lottoNumber -> lottoNumber.equals(LottoNumber.of(bonusNumber)));
+    }
+
+    boolean isDirect() {
+        return this.isDirect;
     }
 
     @Override
@@ -39,11 +41,16 @@ public class Lotto {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         Lotto lotto = (Lotto) o;
-        return Objects.equals(numbers, lotto.numbers);
+        return Objects.equals(lottoNumbers, lotto.lottoNumbers);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(numbers);
+        return Objects.hash(lottoNumbers);
+    }
+
+    @Override
+    public String toString() {
+        return String.join(", ", lottoNumbers.toString());
     }
 }
