+++ src/main/java/lotto/domain/Lottos.java
@@ -1,4 +1,4 @@
-package lotto.domain2;
+package lotto.domain;
 
 import java.util.Collections;
 import java.util.List;
