--- src/main/java/lotto/domain/NumberGenerator.java
@@ -1,8 +1,0 @@
-package lotto.domain;
-
-import java.util.List;
-
-public interface NumberGenerator {
-
-    List<Integer> generate();
-}
