+++ src/main/java/lotto/domain/Rank.java
@@ -1,49 +1,41 @@
 package lotto.domain;
 
 import java.util.Arrays;
-import java.util.function.Predicate;
 
 public enum Rank {
-
     FIRST(6, 2_000_000_000),
-    SECOND(5, true, 30_000_000),
+    SECOND(5, 30_000_000),
     THIRD(5, 1_500_000),
     FOURTH(4, 50_000),
     FIFTH(3, 5_000),
-    MISS(0, 0);
+    LOSE(0, 0);
 
     private final int countOfMatch;
-    private final boolean isMatchBonus;
-    private final int winningMoney;
+    private final int money;
 
-    Rank(final int countOfMatch, final int winningMoney) {
+    Rank(final int countOfMatch, final int money) {
         this.countOfMatch = countOfMatch;
-        this.isMatchBonus = false;
-        this.winningMoney = winningMoney;
+        this.money = money;
     }
 
-    Rank(final int countOfMatch, final boolean isMatchBonus, final int winningMoney) {
-        this.countOfMatch = countOfMatch;
-        this.isMatchBonus = isMatchBonus;
-        this.winningMoney = winningMoney;
-    }
-
-    public int getWinningMoney() {
-        return this.winningMoney;
-    }
-
-    public int getCountOfMatch() {
-        return this.countOfMatch;
+    public int getMoney() {
+        return this.money;
     }
 
     public static Rank findByCountOfMatch(final int countOfMatch, final boolean isMatchBonus) {
-        return Arrays.stream(values())
-                .filter(equals(countOfMatch, isMatchBonus))
+        Rank winningRank = Arrays.stream(values())
+                .filter(rank -> rank.countOfMatch == countOfMatch)
                 .findAny()
-                .orElse(Rank.MISS);
-    }
+                .orElse(Rank.LOSE);
 
-    private static Predicate<? super Rank> equals(final int countOfMatch, final boolean isMatchBonus) {
-        return rank -> rank.countOfMatch == countOfMatch && rank.isMatchBonus == isMatchBonus;
+        if (winningRank == Rank.SECOND && !isMatchBonus) {
+            return Rank.THIRD;
+        }
+
+        if (winningRank == Rank.FOURTH && isMatchBonus) {
+            return Rank.SECOND;
+        }
+
+        return winningRank;
     }
 }
