--- src/main/java/lotto/domain2/LottoMachine.java
@@ -1,7 +1,0 @@
-package lotto.domain2;
-
-import java.util.List;
-
-public interface LottoMachine {
-    Lottos issue(final int countOfTotalLotto, final List<String> directLottos);
-}
