+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,39 +1,43 @@
 package lotto.domain;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
+import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 public class LottoTest {
 
-    private Lotto lotto;
-
-    @BeforeEach
-    void setUp() {
-        lotto = new Lotto(Arrays.asList(1, 2, 3, 4, 5, 6));
-    }
-
     @Test
     void 생성() {
-        assertThat(lotto).isEqualTo(new Lotto(Arrays.asList(1, 2, 3, 4, 5, 6)));
+        List<LottoNumber> lottoNumbers = Arrays.asList(
+                LottoNumber.of(1),
+                LottoNumber.of(2),
+                LottoNumber.of(3),
+                LottoNumber.of(4),
+                LottoNumber.of(5),
+                LottoNumber.of(6)
+        );
+        Lotto lotto = new Lotto(lottoNumbers);
+        assertThat(lotto).isEqualTo(new Lotto(lottoNumbers));
     }
 
     @ParameterizedTest
-    @CsvSource(value = {"1,2,3,8,9,10:3", "1,2,3,4,9,10:4", "1,2,3,4,5,10:5", "1,2,3,4,5,6:6"}, delimiter = ':')
-    void 당첨여부확인(final String input, final int expected) {
-        String[] numbers = input.split(",");
-        List<Integer> winNumbers = new ArrayList<>();
-        for (String number : numbers) {
-            winNumbers.add(Integer.parseInt(number));
+    @ValueSource(strings = {"1,2,3,4,5", "1,2,3,4,5,6,7"})
+    void 로또번호_개수_부족하거나_많을때(final String input) {
+        String[] inputSplit = input.split(",");
+        List<LottoNumber> lottoNumbers = new ArrayList<>();
+        for (String number : inputSplit) {
+            lottoNumbers.add(LottoNumber.of(Integer.parseInt(number)));
         }
-        Rank rank = lotto.checkRank(winNumbers, 45);
-        assertThat(rank.getCountOfMatch()).isEqualTo(expected);
+
+        assertThatThrownBy(() -> {
+            new Lotto(lottoNumbers);
+        }).isInstanceOf(IllegalArgumentException.class);
     }
 }
